+++ src/julia/elements/book1/def_019a_trilateral.jl
@@ -47,7 +47,8 @@ const PhaseDrawSide3 = 3f0
 const PhaseRise = 4f0
 
 
-function random_triangle_point(a::Vector{Float32}, b::Vector{Float32}, c::Vector{Float32})
+function random_triangle_point(
+    a::Vector{Float32}, b::Vector{Float32}, c::Vector{Float32})
     u = rand(Float32)
     v = rand(Float32)
 
@@ -86,17 +87,25 @@ Rectilineal figures are those which are contained by straight lines, trilateral 
 end
 
 function reset_cycle_state(state_ptr::Ptr{Cvoid})
-    line1_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine1HostId))
-    line1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine1Joint2Id))
+    line1_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine1HostId))
+    line1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine1Joint2Id))
 
-    line2_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine2HostId))
-    line2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine2Joint2Id))
+    line2_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine2HostId))
+    line2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine2Joint2Id))
 
-    line3_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine3HostId))
-    line3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine3Joint2Id))
-    triangle_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaTriangleHostId))
+    line3_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine3HostId))
+    line3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine3Joint2Id))
+    triangle_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaTriangleHostId))
 
-    OdinJuliaBridge.hide_point_batch(state_ptr, [line1_host_id, line2_host_id, line3_host_id, triangle_host_id])
+    OdinJuliaBridge.hide_point_batch(state_ptr, [
+        line1_host_id, line2_host_id, line3_host_id, triangle_host_id])
     set_triangle_alpha(state_ptr, triangle_host_id, 0f0)
 
     OdinJuliaBridge.set_point_position(
@@ -139,22 +148,26 @@ function initialize(state_ptr::Ptr{Cvoid})
         VertexC[1], VertexC[2], VertexC[3],
         TriangleColor)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine1HostId, Float32(line1.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine1Joint1Id, Float32(line1.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine1Joint2Id, Float32(line1.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine1HostId, line1.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine1Joint1Id, line1.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine1Joint2Id, line1.joint2_id)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine2HostId, Float32(line2.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine2Joint1Id, Float32(line2.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine2Joint2Id, Float32(line2.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine2HostId, line2.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine2Joint1Id, line2.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine2Joint2Id, line2.joint2_id)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine3HostId, Float32(line3.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine3Joint1Id, Float32(line3.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine3Joint2Id, Float32(line3.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine3HostId, line3.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine3Joint1Id, line3.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaLine3Joint2Id, line3.joint2_id)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaTriangleHostId, Float32(triangle.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaTriangleJoint1Id, Float32(triangle.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaTriangleJoint2Id, Float32(triangle.joint2_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaTriangleJoint3Id, Float32(triangle.joint3_id))
+    OdinJuliaBridge.set_animation_meta(
+        state_ptr, MetaTriangleHostId, triangle.host_id)
+    OdinJuliaBridge.set_animation_meta(
+        state_ptr, MetaTriangleJoint1Id, triangle.joint1_id)
+    OdinJuliaBridge.set_animation_meta(
+        state_ptr, MetaTriangleJoint2Id, triangle.joint2_id)
+    OdinJuliaBridge.set_animation_meta(
+        state_ptr, MetaTriangleJoint3Id, triangle.joint3_id)
 
     reset_cycle_state(state_ptr)
 end
@@ -163,18 +176,28 @@ function clean(state_ptr::Ptr{Cvoid})
 end
 
 function loop(state_ptr::Ptr{Cvoid}, dt::Float32)
-    line1_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine1HostId))
-    line1_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine1Joint1Id))
-    line1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine1Joint2Id))
+    line1_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine1HostId))
+    line1_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine1Joint1Id))
+    line1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine1Joint2Id))
 
-    line2_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine2HostId))
-    line2_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine2Joint1Id))
-    line2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine2Joint2Id))
+    line2_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine2HostId))
+    line2_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine2Joint1Id))
+    line2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine2Joint2Id))
 
-    line3_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine3HostId))
-    line3_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine3Joint1Id))
-    line3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaLine3Joint2Id))
-    triangle_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaTriangleHostId))
+    line3_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine3HostId))
+    line3_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine3Joint1Id))
+    line3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaLine3Joint2Id))
+    triangle_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaTriangleHostId))
 
     if line1_host_id < 0
         return
@@ -195,7 +218,8 @@ function loop(state_ptr::Ptr{Cvoid}, dt::Float32)
     elseif phase == PhaseDrawSide1
         EuclidAnimations.animate_draw_line(
             state_ptr, timer, DrawDuration, VertexA, VertexB,
-            TriangleMaxBrush, TriangleColor, line1_host_id, line1_joint1_id, line1_joint2_id)
+            TriangleMaxBrush, TriangleColor, line1_host_id,
+            line1_joint1_id, line1_joint2_id)
 
         timer += dt
         if timer >= DrawDuration
@@ -205,7 +229,8 @@ function loop(state_ptr::Ptr{Cvoid}, dt::Float32)
     elseif phase == PhaseDrawSide2
         EuclidAnimations.animate_draw_line(
             state_ptr, timer, DrawDuration, VertexB, VertexC,
-            TriangleMaxBrush, TriangleColor, line2_host_id, line2_joint1_id, line2_joint2_id)
+            TriangleMaxBrush, TriangleColor, line2_host_id,
+            line2_joint1_id, line2_joint2_id)
 
         timer += dt
         if timer >= DrawDuration
@@ -215,7 +240,8 @@ function loop(state_ptr::Ptr{Cvoid}, dt::Float32)
     elseif phase == PhaseDrawSide3
         EuclidAnimations.animate_draw_line(
             state_ptr, timer, DrawDuration, VertexC, VertexA,
-            TriangleMaxBrush, TriangleColor, line3_host_id, line3_joint1_id, line3_joint2_id)
+            TriangleMaxBrush, TriangleColor, line3_host_id,
+            line3_joint1_id, line3_joint2_id)
 
         timer += dt
         if timer >= DrawDuration
